requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
tavily-python
